Remove dangerous ST hack for recent base library
GHC 8.0 added a `runRW#` primop to work around the issue that led
to the unsafe reimplementation of `runST` in `unordered-containers`.
Use the standard `runST` for `base` 4.9.0 and above.

This commit *does not* remove the `-fno-full-laziness` invocations
in various modules; doing so will require testing to ensure that it
is good (or at least not bad) for performance.