Traverse arrays better
The previous approach was to convert the array to a list, traverse
the list, and then put it back into an array. I figured it might
be worth trying a more direct approach. It seems to work
rather better. I also added rules to use a very direct implementation
for `ST` and `IO`.