Fix infinite loop in `isSubmapOf[By]` / `isSubsetOf` on 32-bit platforms (#501)
The internal `submapBitmapIndexed` function used by these functions could enter an infinite loop while comparing two nodes `m1` and `m2` where `m2` contained a sub-node associated with the partial hash `0b11111`. In this case the high bit of the combined bitmap of `m1` and `m2`, `b1Orb2`, was `1`. After checking the submap condition at the high bit, the 32-bit `m` variable used to iterate over `b1Orb2` would be left-shifted and overflow to `0`, resulting in the infinite loop. 

https://github.com/haskell-unordered-containers/unordered-containers/blob/37eee2290cab287c0947d1de20235a37ced63c94/Data/HashMap/Internal.hs#L1551-L1572

To fix this bug and any similar issues related to a branching factor of 32, we return to a branching factor of 16 for 32-bit platforms. On 64-bit platforms the branching factors stays at 32.

Fixes #491.